complete task handling done
When rendering a set of task cards, the task object's isComplete field is
checked and all completed tasks are put below uncompleted tasks.

The completed task gets a green background and the text a grey color.

the checkbox toggles the task object's isComplete field and moves the task
card to the bottom